+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/CustomDataProviderTest.kt
@@ -55,7 +55,7 @@ internal class CustomDataProviderTest : BaseTest() {
 
         mockServer = MockWebServer()
 
-        MapboxSearchSdk.reinitializeInternal(
+        MapboxSearchSdk.initializeInternal(
             application = targetApplication,
             timeProvider = timeProvider,
             keyboardLocaleProvider = keyboardLocaleProvider,