+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/ReverseGeocodingSearchIntegrationTest.kt
@@ -73,7 +73,7 @@ internal class ReverseGeocodingSearchIntegrationTest : BaseTest() {
 
         mockServer = MockWebServer()
 
-        MapboxSearchSdk.reinitializeInternal(
+        MapboxSearchSdk.initializeInternal(
             application = targetApplication,
             timeProvider = timeProvider,
             keyboardLocaleProvider = keyboardLocaleProvider,