+++ CHANGELOG.md
@@ -9,9 +9,12 @@
 - [CORE] `ServiceProvider.locationEngine()` function has been removed. Now `LocationEngine` instance can be retrieved from settings used for search engines instantiation.
 - [CORE] `SearchSdkSettings` class has been removed. Now default max number of history records is always 100.
 - [CORE] Now `SearchEngineSettings` and `OfflineSearchEngineSettings` accept new constructor parameters. `OfflineSearchEngineSettings.copy()` function's signature and `OfflineSearchEngineSettings.Builder` have also been changed. `SearchEngineSettings` provides `SearchEngineSettings.Builder` type and new `copy()` and `toBuilder()` functions.
+- [CORE] `MapboxSearchSdk.initialize()` function has been removed. Now it's not needed to initialize Search SDK explicitly.
 - [CORE] `MapboxSearchSdk.setAccessToken()` function has been removed. Now in case of changed token a new instance of `SearchEngine` or `OfflineSearchEngine` should be created.
-- [CORE] `MapboxSearchSdk.getSearchEngine()` function has been removed. Now call `MapboxSearchSdk.createSearchEngine()` function instead.
+- [CORE] `MapboxSearchSdk.getSearchEngine()` function has been removed, call `MapboxSearchSdk.createSearchEngine()` function instead.
 - [CORE] `MapboxSearchSdk.createSearchEngine()` function without parameters has been removed. Now `SearchEngineSettings` should always be provided as an argument.
+- [CORE] `MapboxSearchSdk.getOfflineSearchEngine()` function has been removed, call `MapboxSearchSdk.createOfflineSearchEngine()` function instead.
+- [CORE] `OfflineSearchEngine.tileStore` property has been remove, call `OfflineSearchEngine.settings.tileStore` instead.
 - [UI] `SearchResultsView.initialize()` now accepts `SearchResultsView.Configuration` object.
 
 ### New features